+++ examples/1-getting-started/src/app/index.tsx
@@ -1,11 +1,13 @@
 import React, { FC } from 'react';
 import * as ReactDOM from 'react-dom';
-import { createClient, Provider } from 'urql';
+import { createClient, Provider, defaultExchanges } from 'urql';
+import { devtoolsExchange } from '@urql/devtools';
 import { Home } from './Home';
 import './index.css';
 
 const client = createClient({
   url: 'http://localhost:3001/graphql',
+  exchanges: [devtoolsExchange, ...defaultExchanges],
 });
 
 export const App: FC = () => (