+++ src/hooks/useMutation.test.tsx
@@ -6,7 +6,7 @@ jest.mock('../client', () => {
   const mock = {
     executeMutation: jest.fn(() =>
       pipe(
-        fromValue({ data: 1, error: 2 }),
+        fromValue({ data: 1, error: 2, extensions: { i: 1 } }),
         delay(200)
       )
     ),
@@ -116,6 +116,14 @@ describe('on execute', () => {
       source: 'MutationUser',
     });
   });
+
+  it('can adjust context in executeMutation', () => {
+    renderer.create(<MutationUser {...props} />);
+    act(() => {
+      execute(vars, { url: 'test' });
+    });
+    expect(client.executeMutation.mock.calls[0][1].url).toBe('test');
+  });
 });
 
 describe('on subscription update', () => {
@@ -133,6 +141,14 @@ describe('on subscription update', () => {
     wrapper.update(<MutationUser {...props} />);
 
     expect(state).toHaveProperty('error', 2);
+  });
+
+  it('forwards extensions response', async () => {
+    const wrapper = renderer.create(<MutationUser {...props} />);
+    await execute();
+    wrapper.update(<MutationUser {...props} />);
+
+    expect(state).toHaveProperty('extensions', { i: 1 });
   });
 
   it('sets fetching to false', async () => {