+++ src/hooks/useMutation.ts
@@ -2,7 +2,7 @@ import { DocumentNode } from 'graphql';
 import { useContext, useCallback } from 'react';
 import { pipe, toPromise } from 'wonka';
 import { Context } from '../context';
-import { OperationResult } from '../types';
+import { OperationResult, OperationContext } from '../types';
 import { CombinedError, createRequest } from '../utils';
 import { useImmediateState } from './useImmediateState';
 import { useDevtoolsContext } from './useDevtoolsContext';
@@ -11,6 +11,7 @@ export interface UseMutationState<T> {
   fetching: boolean;
   data?: T;
   error?: CombinedError;
+  extensions?: Record<string, any>;
 }
 
 export type UseMutationResponse<T, V> = [
@@ -27,20 +28,26 @@ export const useMutation = <T = any, V = object>(
     fetching: false,
     error: undefined,
     data: undefined,
+    extensions: undefined,
   });
 
   const executeMutation = useCallback(
-    (variables?: V) => {
-      setState({ fetching: true, error: undefined, data: undefined });
+    (variables?: V, context?: Partial<OperationContext>) => {
+      setState({
+        fetching: true,
+        error: undefined,
+        data: undefined,
+        extensions: undefined,
+      });
 
       const request = createRequest(query, variables as any);
 
       return pipe(
-        client.executeMutation(request, devtoolsContext),
+        client.executeMutation(request, { ...devtoolsContext, ...context }),
         toPromise
       ).then(result => {
-        const { data, error } = result;
-        setState({ fetching: false, data, error });
+        const { data, error, extensions } = result;
+        setState({ fetching: false, data, error, extensions });
         return result;
       });
     },