+++ src/hooks/useQuery.ts
@@ -13,6 +13,7 @@ export interface UseQueryArgs<V> {
   query: string | DocumentNode;
   variables?: V;
   requestPolicy?: RequestPolicy;
+  context?: Partial<OperationContext>;
   pause?: boolean;
 }
 
@@ -20,6 +21,7 @@ export interface UseQueryState<T> {
   fetching: boolean;
   data?: T;
   error?: CombinedError;
+  extensions?: Record<string, any>;
 }
 
 export type UseQueryResponse<T> = [
@@ -40,6 +42,7 @@ export const useQuery = <T = any, V = object>(
     fetching: false,
     data: undefined,
     error: undefined,
+    extensions: undefined,
   });
 
   // This creates a request which will keep a stable reference
@@ -55,15 +58,23 @@ export const useQuery = <T = any, V = object>(
       [unsubscribe.current] = pipe(
         client.executeQuery(request, {
           requestPolicy: args.requestPolicy,
+          ...args.context,
           ...opts,
           ...devtoolsContext,
         }),
-        subscribe(({ data, error }) => {
-          setState({ fetching: false, data, error });
+        subscribe(({ data, error, extensions }) => {
+          setState({ fetching: false, data, error, extensions });
         })
       );
     },
-    [args.requestPolicy, client, devtoolsContext, request, setState]
+    [
+      args.context,
+      args.requestPolicy,
+      client,
+      devtoolsContext,
+      request,
+      setState,
+    ]
   );
 
   useImmediateEffect(() => {