+++ src/hooks/useSubscription.ts
@@ -6,10 +6,12 @@ import { CombinedError, noop } from '../utils';
 import { useDevtoolsContext } from './useDevtoolsContext';
 import { useRequest } from './useRequest';
 import { useImmediateState } from './useImmediateState';
+import { OperationContext } from '../types';
 
 export interface UseSubscriptionArgs<V> {
   query: DocumentNode | string;
   variables?: V;
+  context?: Partial<OperationContext>;
 }
 
 export type SubscriptionHandler<T, R> = (prev: R | undefined, data: T) => R;
@@ -18,6 +20,7 @@ export interface UseSubscriptionState<T> {
   fetching: boolean;
   data?: T;
   error?: CombinedError;
+  extensions?: Record<string, any>;
 }
 
 export type UseSubscriptionResponse<T> = [UseSubscriptionState<T>];
@@ -34,6 +37,7 @@ export const useSubscription = <T = any, R = T, V = object>(
     fetching: true,
     error: undefined,
     data: undefined,
+    extensions: undefined,
   });
 
   // This creates a request which will keep a stable reference
@@ -44,16 +48,20 @@ export const useSubscription = <T = any, R = T, V = object>(
     unsubscribe.current();
 
     [unsubscribe.current] = pipe(
-      client.executeSubscription(request, devtoolsContext),
-      subscribe(({ data, error }) => {
+      client.executeSubscription(request, {
+        ...devtoolsContext,
+        ...args.context,
+      }),
+      subscribe(({ data, error, extensions }) => {
         setState(s => ({
           fetching: true,
           data: handler !== undefined ? handler(s.data, data) : data,
           error,
+          extensions,
         }));
       })
     );
-  }, [client, devtoolsContext, handler, request, setState]);
+  }, [client, devtoolsContext, handler, request, setState, args.context]);
 
   // Trigger subscription on query change
   // We don't use useImmediateEffect here as we have no way of